mlx5: Fix byte_count type in umr_sg_list_create
[ Upstream commit e23cb0a72e454d0f985316679758fe2ff3320988 ]

Change byte_count variable from int to uint64_t to match the
reglen parameter type and prevent potential overflow issues when
working with large memory registrations in UMR (User-Mode Memory
Registration) operations.

The function signature uses uint64_t *reglen, but byte_count was
declared as int, which could cause type mismatch issues and potential
overflow when accumulating large transfer sizes across multiple SGEs.

+++ providers/mlx5/qp.c
@@ -2099,7 +2099,7 @@ static int umr_sg_list_create(struct mlx5_qp *qp,
 			      uint64_t *reglen)
 {
 	struct mlx5_wqe_data_seg *dseg;
-	int byte_count = 0;
+	uint64_t byte_count = 0;
 	int i;
 	size_t tmp;
